% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMP.R
\docType{data}
\name{CMP}
\alias{CMP}
\title{CMP dataset}
\format{
A dataframe with 473 rows and 9 variables:
\describe{
  \item{party}{dbl CMP ID for each political party}
  \item{date}{dbl Date of the election in YYYYMM format} 
  \item{countryname}{chr String denoting the country}
  \item{sdper103}{dbl Denotes score for per103 (mentions of anti-imperialism)}
  \item{wasingov}{dbl Binary variable denoting whether or not that party has been in government since the last election}
  \item{timeingov2}{dbl Denotes how many days that party was in government since the last election}
  \item{decade}{chr Denotes the decade that observation took place in}
  \item{pervote}{dbl Standardised variable representing the percentage vote obtained by that party in that election}
  \item{lag_sd_rile`}{dbl Standardised right-left score (rile) for that party at the previous election}
}
}
\usage{
CMP
}
\description{
Subset of Comparative Manifesto Project database, with additional variables indicating whether or not each party has been in government (data from Parl Gov database and Party Facts)
}
\keyword{datasets}
