% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparse_PanelEstimate.R
\name{Sparse_PanelEstimate}
\alias{Sparse_PanelEstimate}
\title{Estimate DiD estimator for Sparse_PanelMatch object with bootstrapped standard errors}
\usage{
Sparse_PanelEstimate(data, n_iterations = 1000, alpha = 0.05)
}
\arguments{
\item{data}{A SparsePanelMatch}

\item{n_iterations}{The number of iterations to use for bootstrapping (defaults to 1000)}

\item{alpha}{The alpha level to use when calculating the confidence intervals (defaults to 0.05)}
}
\value{
A SparsePanelEstimate object
}
\description{
Estimate DiD estimator for Sparse_PanelMatch object with bootstrapped standard errors
}
\examples{
MatchedData <- Sparse_PanelMatch(data = CMP, time = "date", unit = "party", treatment = "wasingov", outcome = "sdper103", treatment_lags = 3, outcome_leads = 0, time_window_in_months = 60, match_missing = TRUE, covs = c("pervote", "lag_sd_rile"), qoi = "att", refinement_method = "mahalanobis", size_match = 5, use_diagonal_covmat = TRUE)
Sparse_PanelEstimate(data = MatchedData, n_iterations = 1000, alpha = 0.05)
}
