% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparse_PanelMatch.R
\name{Sparse_PanelMatch}
\alias{Sparse_PanelMatch}
\title{Find matching units across irregular panel data:
Stage 1: Match each treated observation (unit-time) with control observations occurring within a user-defined time window (this is the only significant difference to the original method, where it is assumed that the panel data is well-ordered and regular, meaning that each observation is matched with every other observation at that year/month/date)
Stage 2: Within that matched set, find control observations with exactly the same treatment history over the last n period (e.g. over the last 3 election cycles)
Stage 3: Within that matched set, refine further using either Propensity Scores, Covariate Balancing Propensity Scores or Mahalanobis distance}
\usage{
Sparse_PanelMatch(
  data,
  time,
  unit,
  treatment,
  outcome,
  treatment_lags,
  outcome_leads,
  time_window_in_months,
  match_missing = TRUE,
  covs,
  qoi = c("att", "atc"),
  refinement_method = c("none", "CBPS.weight", "CBPS.match", "ps.weight", "ps.match",
    "mahalanobis"),
  size_match,
  use_diagonal_covmat = FALSE
)
}
\arguments{
\item{data}{The dataset}

\item{time}{The name of the time variable, quoted as a string, in YYYYMM format and structured as numeric variable.}

\item{unit}{The name of the unit variable, quoted as a string, structured as a numeric variable.}

\item{treatment}{The name of the treatment variable, quoted as a string, should be numeric variable where 0 = untreated and 1 = treated.}

\item{treatment_lags}{The number of treatment lags to use for the exact matching on treatment history.}

\item{outcome_leads}{The number of outcome leads for which to calculate DiD estimators (each compared to Yt-1) [single integer].}

\item{time_window_in_months}{The size of the time window to search for control observations within [single integer].}

\item{match_missing}{Whether or not to match missing values within treatment history (defaults to TRUE).}

\item{covs}{List of covariates to use in matching refinement procedures (should be a list of variables quoted as strings).}

\item{qoi}{Quantity of interest (either ATT [Average Treatment effect on Treated] or ATC [Average Treatment effect on Controls]).}

\item{refinement_method}{Refinement method to use at Stage 3. Can take one of "none","CBPS.weight" (adjusts weights by Covariate Balancing Propensity Scores), "CBPS.match" (reduces the size of each control set to size_match by lowest CBPS), "ps.weight" and "ps.match" do the same for standard propensity scores, "mahalanobis" reduces set size to size_match by Mahalanobis distance. For more details see Imai, Kim & Wang (2018).}

\item{size_match}{Maximum size of sets. For use with 'CBPS.match', 'ps.match' and 'mahalanobis'.}
}
\value{
A Sparse_PanelMatch object
}
\description{
Find matching units across irregular panel data:
Stage 1: Match each treated observation (unit-time) with control observations occurring within a user-defined time window (this is the only significant difference to the original method, where it is assumed that the panel data is well-ordered and regular, meaning that each observation is matched with every other observation at that year/month/date)
Stage 2: Within that matched set, find control observations with exactly the same treatment history over the last n period (e.g. over the last 3 election cycles)
Stage 3: Within that matched set, refine further using either Propensity Scores, Covariate Balancing Propensity Scores or Mahalanobis distance
}
\examples{
Sparse_PanelMatch(data = cmp, time = "date", unit = "party", treatment = "wasingov", outcome = "sdper103", treatment_lags = 3, outcome_leads = 0, time_window_in_months = 60, match_missing = TRUE, covs = c("pervote", "lag_sd_rile"), qoi = "att", refinement_method = "mahalanobis", size_match = 5, use_diagonal_covmat = TRUE)
}
