% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparse_PanelEstimate.R
\name{summary.SparsePanelEstimate}
\alias{summary.SparsePanelEstimate}
\title{summary.SparsePanelEstimate}
\usage{
\method{summary}{SparsePanelEstimate}(object)
}
\arguments{
\item{object}{A SparsePanelEstimate object}
}
\description{
Summarises a SparsePanelEstimate object. By default returns standard percentile confidence intervals. Bias corrected confidence intervals (Efron & Tibshirani 1993 p138) are also available in the SparsePanelEstimate object.
}
\examples{
MatchedData <- Sparse_PanelMatch(data = CMP, time = "date", unit = "party", treatment = "wasingov", outcome = "sdper103", treatment_lags = 3, outcome_leads = 0, time_window_in_months = 60, match_missing = TRUE, covs = c("pervote", "lag_sd_rile"), qoi = "att", refinement_method = "mahalanobis", size_match = 5, use_diagonal_covmat = TRUE)
Estimate <- Sparse_PanelEstimate(data = MatchedData, n_iterations = 1000, alpha = 0.05)
summary(Estimate, bias_correction = TRUE)
}
