% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparse_PanelEstimate.R
\name{Sparse_PanelEstimate}
\alias{Sparse_PanelEstimate}
\title{Sparse_PanelEstimate}
\usage{
Sparse_PanelEstimate(data, n_iterations = 1000, alpha = 0.05)
}
\arguments{
\item{data}{A SparsePanelMatch}

\item{n_iterations}{The number of iterations to use for bootstrapping (defaults to 1000)}

\item{alpha}{The alpha level to use when calculating the confidence intervals (defaults to 0.05)}
}
\value{
A SparsePanelEstimate object
}
\description{
Estimate DiD estimator for Sparse_PanelMatch object with bootstrapped standard errors
}
\examples{
Sparse_PanelEstimate(data = SparsePanelMatch1, n_iterations = 1000, alpha = 0.05)
}
