% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparse_PanelMatch.R
\name{Sparse_PanelMatch}
\alias{Sparse_PanelMatch}
\title{Find matching units across irregular panel data:
1. Match each treated observation (unit-time) to untreated observations occurring within a user-defined time window 
2. Limit that matched set to control observations with exactly the same treatment history over the last n observations
3. Within that matched set, refine further or weight using Propensity Scores, Covariate Balancing Propensity Scores or Mahalanobis distance}
\usage{
Sparse_PanelMatch(
  data,
  time,
  unit,
  treatment,
  outcome,
  treatment_lags,
  outcome_leads,
  time_window_in_months,
  match_missing = TRUE,
  covs,
  qoi = c("att", "atc"),
  refinement_method = c("none", "CBPS.weight", "CBPS.match", "ps.weight", "ps.match",
    "mahalanobis"),
  size_match,
  use_diagonal_covmat = FALSE
)
}
\arguments{
\item{data}{The name of the dataset}

\item{time}{String of characters representing the name of the time variable. Should be in YYYYMM (e.g. 199204) format and structured as numeric variable.}

\item{unit}{String of characters representing the name of the unit variable. Should be a numeric variable.}

\item{treatment}{String of characters representing the name of the treatment variable. Should be numeric variable where 0 = untreated and 1 = treated.}

\item{outcome}{String of characters representing the name of the outcome variable. Should be a numeric variable.}

\item{treatment_lags}{Integer representing the number of treatment lags to use for the exact matching on treatment history.}

\item{outcome_leads}{Integer representing the number of outcome leads for which to calculate DiD estimands (each compared to Yt-1).}

\item{time_window_in_months}{Integer representing the size of the time window in months within which to search for control observations.}

\item{match_missing}{Whether or not to match missing values within treatment history (defaults to TRUE).}

\item{covs}{List of covariates to use in matching refinement procedures (should be a list of strings of characters).}

\item{qoi}{The wuantity of interest: either 'ATT' (Average Treatment effect on Treated) or 'ATC' (Average Treatment effect on Controls).}

\item{refinement_method}{The refinement method used to improve covariate balancing after the exact matching procedure. Can take one of: "none"; "CBPS.weight" (adjusts weights by Covariate Balancing Propensity Scores); "CBPS.match" (reduces the size of each control set to size_match by lowest CBPS); "ps.weight" and "ps.match" do the same for standard propensity scores; "mahalanobis" reduces set size to size_match by Mahalanobis distance. For more details see Imai, Kim & Wang (2018).}

\item{size_match}{Integer representing the maximum number of control observations within each refined set. For use with 'CBPS.match', 'ps.match' and 'mahalanobis'. Has no impact otherwise.}

\item{use_diagonal_covmat}{Whether or not mahalanobis distance calculation should use a regular covariance matrix (FALSE), or a diagonal matrix with only covariate variances (TRUE). Imai, Kim & Wang (2018) report that, "in many cases, setting this to TRUE can lead to better covariate balance, especially when there is high correlation between variables". Default is FALSE.}
}
\value{
A Sparse_PanelMatch object
}
\description{
Find matching units across irregular panel data:
1. Match each treated observation (unit-time) to untreated observations occurring within a user-defined time window 
2. Limit that matched set to control observations with exactly the same treatment history over the last n observations
3. Within that matched set, refine further or weight using Propensity Scores, Covariate Balancing Propensity Scores or Mahalanobis distance
}
\examples{
MatchedData <- Sparse_PanelMatch(data = CMP, time = "date", unit = "party", treatment = "wasingov", outcome = "sdper103", treatment_lags = 3, outcome_leads = 0, time_window_in_months = 60, match_missing = TRUE, covs = c("pervote", "lag_sd_rile"), qoi = "att", refinement_method = "mahalanobis", size_match = 5, use_diagonal_covmat = TRUE)
}
