% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparse_PanelEstimate.R
\name{summary.SparsePanelEstimate}
\alias{summary.SparsePanelEstimate}
\title{Summarises a SparsePanelEstimate object}
\usage{
\method{summary}{SparsePanelEstimate}(data, bias_correction = FALSE)
}
\arguments{
\item{data}{A SparsePanelEstimate object}

\item{bias_correction}{Whether to display bias corrected confidence intervals (Efron & Tibshirani 1993 p138), or standard percentile confidence intervals, Defaults to FALSE (standard percentiles).}
}
\description{
Summarises a SparsePanelEstimate object
}
\examples{
summary(SparsePanelEstimate1)
}
